Add hook to remove `get_term_adjust_id` filter
Adds a hook to remove the WPML `get_term_adjust_id` filter that is
attached to the WP `get_term` function at a very high priority. This
filter causes WPML to adjust term ids for every request - shifting
request terms to the "current language". The problem is that the
WPGraphQL requests don't typically use the current language (and
especially not for use cases like Gatsby). This results in term IDs only
ever being returned for the default language and with single term
queries always returning the default language's term.

The function is added to the `init_graphql_request` action that runs
at the beginning of the GraphQL request cycle. Using this action
ensures that we're making adjustments for only "graphql" requests and
that those changes are early enough in the cycle to get in front of
the WPML filters.