Add slug field to term translations
Adds the `slug` field to the term translations response (type and
resolver) to help when a term may have the same slug between different
locales. When the same slug is used, WordPress will add a numerical
suffix to the href/link even though the same slug may be desired.

Adding the slug to the response allows the client to use the slug when
generating routes. For example, imagine the situation with two locales:
English and Spanish. Creating a `tech` tag in both locales would look
like this:

  * English
    - slug: `tech`
    - href: `https://.../tag/tech`

  * Spanish
    - slug: `tech`
    - href: `https://.../es/tag/tech-2`

With the slug and the href, the client can use whatever works best in
that context.